+++ BrightFutures/Promise.swift
@@ -57,7 +57,7 @@ public class Promise<T> {
         return self.future.tryError(error)
     }
     
-    public func tryComplete(result: TaskResult<T>) -> Bool {
+    public func tryComplete(result: Result<T>) -> Bool {
         return self.future.tryComplete(result)
     }