+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# Unreleased changes
+- `TaskResultValueWrapper` has been renamed to the conventional name `Box`
+- `TaskResult` has been renamed to the conventional name `Result`
+
 # 1.0.0-beta.1
 This release marks the state of the project before this changelog was kept up to date.